fix(types): play global type error
