+++ tsconfig.web.json
@@ -7,7 +7,11 @@
     "types": ["unplugin-vue-macros/macros-global"],
     "skipLibCheck": true
   },
-  "include": ["packages", "typings/env.d.ts"],
+  "include": [
+    "packages/utils/vue3.3.polyfill.ts",
+    "packages",
+    "typings/env.d.ts"
+  ],
   "exclude": [
     "node_modules",
     "**/dist",
